Moved enum subclasses into fpdf.enums
